address PR comments: add AbstractIndexInsightTask and remove IndexInsightUtils
