+++ plugin/src/test/java/org/opensearch/ml/action/IndexInsight/GetIndexInsightTransportActionTests.java
@@ -22,7 +22,6 @@ import java.time.Instant;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.CompletionStage;
 
 import org.junit.Before;
 import org.junit.Test;
